Collapse the /lib vs /python source code split
The code is currently split into pure C code (/lib) and the Python
bindings (/python). This sounds nice until you actually look at it and
realise just how much this separation costs.

* Most exceptions are OverflowError because the code that raises
  exceptions can't reference true exception types provided by Python.h

* Python functions have to be wrapped in trivial callback functions
  passed around as void pointers in structs so that the encode/decoder
  can avoid directly referencing the Python functions

* void pointer anarchy

Now:

  lib/ultrajsondec.c + python/JSONtoObj.c -> decode.c
  lib/ultrajsonenc.c + python/objToJSON.c -> encode.c

Changes beyond the move/concatenation are as minimal as I could make it
(deduplicating #includes, adjusting paths, shuffling function prototypes
around).
